allow null error attributes
some server implementations send null values for `source` etc. this
results in corresponding `nil` values in the error `attrs` hash.

unfortunately `fetch` does not return the default value when the hash
value is `nil`. so when it returns `nil` for `source` we get the
exception: `undefined method `fetch' for nil:NilClass` when trying to
treat it as a hash.

the solution is to use `attrs[:source] || {}` instead of
`attrs.fetch(:source, {})`. this is also necessary for a few other
error attributes.