Update bundler before in Travis CI builds
Builds are broken because the `bundler` version that Travis CI
automatically uses at this time (`1.7.6`) is coming across an
error that was fixed in a later version/release.

Having an outdated version of `bundler` in the Travis CI builds is an
ongoing issue and this is currently the easiest work-around.

See:

* bundler/bundler#3558
* travis-ci/travis-ci#3531