Rework Errors::NotFound
Make the constructor consistent with parent class (Errors::ApiError)
so we can pass along a Rack/Faraday environment (if provided) and take
advantage of the JSON API error message handling implemented on parent
class.

We do have a case where we're explicitly raising an Errors::NotFound
outside the status middleware. This occurs in the Query::Builder class.
In this case, we're passing nil for the env and an explicit error message.